refactor @parties
