+++ src/components/@parties/MyPartiesScene.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { array, object, bool } from 'prop-types'
 import { withStyles } from '@material-ui/core/styles'
-import PartiesCard from './PartiesCard'
+import PartiesList from './PartiesList'
 import isEmpty from 'lodash/isEmpty'
 import Loading from '../Loading'
 import NotFound from '../NotFound'
@@ -16,7 +16,12 @@ const styles = {
 
 class MyPartiesScene extends React.Component {
   componentWillMount() {
-    this.props.actions.parties.userLoad(this.props.match.params.id)
+    const { actions, match } = this.props
+    actions.parties.userLoad(match.params.id)
+  }
+
+  like = (id) => {
+    this.props.actions.like.like(id)
   }
 
   render() {
@@ -26,7 +31,7 @@ class MyPartiesScene extends React.Component {
 
     return (
       <div className={classes.root}>
-        <PartiesCard />
+        <PartiesList onLike={this.like} parties={parties} />
       </div>
     )
   }
@@ -34,13 +39,9 @@ class MyPartiesScene extends React.Component {
 
 MyPartiesScene.propTypes = {
   classes: object.isRequired,
-  parties: array,
+  parties: array.isRequired,
   loading: bool.isRequired,
   actions: object.isRequired,
   match: object.isRequired,
-}
-
-MyPartiesScene.defaultProps = {
-  parties: [],
 }
 export default withStyles(styles)(connector(MyPartiesScene))
